refactor: extract getOrderedItems utility and memoize sensors
